+++ llp/lib_io.h
@@ -21,7 +21,7 @@ typedef enum _e_ot{
 int sl_Rstr(slice* sl, char** out);
 int sl_Rlens(slice* sl, unsigned int* out);
 int sl_Rbyte(slice* sl, byte* out);
-int sl_Ruint(slice* sl, size_t* out);
+int sl_Ruint(slice* sl, unsigned int* out);
 
 int llp_out_close(slice* out);
 int llp_out_clr(slice* out);
